Moved GitLabLink to helpers package
